+++ packages/table/src/components/EditableTable/index.tsx
@@ -9,7 +9,8 @@ import { Field } from 'rc-field-form';
 import ProTable from '../../Table';
 import type { ProTableProps, ActionType } from '../../typing';
 import type { GetRowKey } from 'antd/lib/table/interface';
-import { ProFormDependency, ProFormInstance } from '@ant-design/pro-form';
+import type { ProFormInstance } from '@ant-design/pro-form';
+import { ProFormDependency } from '@ant-design/pro-form';
 import get from 'rc-util/lib/utils/get';
 
 export type RecordCreatorProps<DataSourceType> = {
@@ -112,7 +113,7 @@ function EditableTable<
     onChange: props.onChange,
   });
   const getRowKey = React.useMemo<GetRowKey<DataType>>((): GetRowKey<DataType> => {
-    if (typeof rowKey === 'function' && rowKey) {
+    if (typeof rowKey === 'function') {
       return rowKey;
     }
     return (record: DataType, index?: number) => (record as any)[rowKey as string] || index;
@@ -260,17 +261,19 @@ function EditableTable<
         />
       </EditableTableActionContext.Provider>
       {/* 模拟 onValuesChange */}
-      <ProFormDependency name={[props.name!]}>
-        {(changeValue) => {
-          const list = get(changeValue, [props.name].flat(1) as string[]) as any[];
-          const changeItem = list?.find((item, index) => {
-            return !isDeepEqualReact(item, preData?.[index]);
-          });
-          if (!changeItem) return null;
-          props?.editable?.onValuesChange?.(changeItem, list);
-          return null;
-        }}
-      </ProFormDependency>
+      {props.name ? (
+        <ProFormDependency name={[props.name!]}>
+          {(changeValue) => {
+            const list = get(changeValue, [props.name].flat(1) as string[]) as any[];
+            const changeItem = list?.find((item, index) => {
+              return !isDeepEqualReact(item, preData?.[index]);
+            });
+            if (!changeItem) return null;
+            props?.editable?.onValuesChange?.(changeItem, list);
+            return null;
+          }}
+        </ProFormDependency>
+      ) : null}
     </>
   );
 }
@@ -280,15 +283,13 @@ function FieldEditableTable<
   Params extends ParamsType = ParamsType,
   ValueType = 'text',
 >(props: EditableProTableProps<DataType, Params, ValueType>) {
-  const { name, formItemProps } = props;
-
-  if (!name) return <EditableTable<DataType, Params, ValueType> {...props} />;
+  if (!props.name) return <EditableTable<DataType, Params, ValueType> {...props} />;
   return (
     <Form.Item
       style={{
         maxWidth: '100%',
       }}
-      {...formItemProps}
+      {...props?.formItemProps}
       name={props.name}
     >
       <>