+++ tests/form/base.test.tsx
@@ -1,6 +1,7 @@
 import React, { useRef, useEffect } from 'react';
 import { Button, Input } from 'antd';
 import type { ProFormInstance } from '@ant-design/pro-form';
+import { ProFormDateTimePicker } from '@ant-design/pro-form';
 import { ProFormDigitRange } from '@ant-design/pro-form';
 import ProForm, {
   ProFormText,
@@ -15,6 +16,7 @@ import { act } from 'react-dom/test-utils';
 import { FontSizeOutlined } from '@ant-design/icons';
 import { mount } from 'enzyme';
 import { waitTime, waitForComponentToPaint } from '../util';
+import { render as reactRender } from '@testing-library/react';
 import moment from 'moment';
 
 describe('ProForm', () => {
@@ -1896,5 +1898,83 @@ describe('ProForm', () => {
     await waitForComponentToPaint(wrapper, 100);
 
     expect(onFinish).toBeCalledWith(undefined);
+  });
+
+  it('📦 when dateFormatter is a Function', async () => {
+    const fn1 = jest.fn();
+    const fn2 = jest.fn();
+    const App = () => {
+      return (
+        <ProForm
+          dateFormatter={(value, valueType) => {
+            fn1(value.format('YYYY/MM/DD HH:mm:ss'), valueType);
+            return value.format('YYYY/MM/DD HH:mm:ss');
+          }}
+          onFinish={async (values) => {
+            fn2(values.datetime);
+            return true;
+          }}
+        >
+          <ProFormDateTimePicker
+            name="datetime"
+            initialValue={moment('2021-08-09 12:12:12')}
+            fieldProps={{ open: true }}
+          />
+        </ProForm>
+      );
+    };
+
+    const wrapper = mount(<App />);
+    await waitForComponentToPaint(wrapper);
+
+    expect(fn1).toBeCalledWith('2021/08/09 12:12:12', 'dateTime');
+    act(() => {
+      wrapper.find('button.ant-btn-primary').at(1).simulate('click');
+    });
+
+    await waitForComponentToPaint(wrapper, 100);
+    expect(fn2).toHaveBeenCalledWith('2021/08/09 12:12:12');
+
+    act(() => {
+      expect(wrapper.render()).toMatchSnapshot();
+    });
+  });
+
+  it(`📦 rules change should rerender`, () => {
+    const html = reactRender(
+      <ProForm>
+        <ProFormText
+          width="md"
+          rules={[
+            {
+              required: true,
+              message: 'test',
+            },
+          ]}
+          name="function"
+          label="生效方式"
+        />
+      </ProForm>,
+    );
+
+    expect(html.baseElement.querySelectorAll('.ant-form-item-required').length).toBe(1);
+
+    html.rerender(
+      <ProForm>
+        <ProFormText
+          width="md"
+          rules={[
+            {
+              required: false,
+              message: 'test',
+            },
+          ]}
+          name="function"
+          label="生效方式"
+        />
+      </ProForm>,
+    );
+
+    expect(html.baseElement.querySelectorAll('.ant-form-item-required').length).toBe(0);
   });
 });