+++ packages/list/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.21.30](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.29...@ant-design/pro-list@1.21.30) (2022-02-21)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.29](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.28...@ant-design/pro-list@1.21.29) (2022-02-18)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
+## [1.21.28](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.27...@ant-design/pro-list@1.21.28) (2022-02-15)
+
+**Note:** Version bump only for package @ant-design/pro-list
+
 ## [1.21.27](https://github.com/ant-design/pro-components/compare/@ant-design/pro-list@1.21.26...@ant-design/pro-list@1.21.27) (2022-02-14)
 
 **Note:** Version bump only for package @ant-design/pro-list